Refactor client into modular components
