+++ include/lantern/core/client/options.h
@@ -1,0 +1,54 @@
+#ifndef LANTERN_CORE_CLIENT_OPTIONS_H
+#define LANTERN_CORE_CLIENT_OPTIONS_H
+
+#include <stdint.h>
+
+#include "lantern/support/string_list.h"
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+#define LANTERN_DEFAULT_DATA_DIR "./data"
+#define LANTERN_DEFAULT_GENESIS_CONFIG "./genesis/config.yaml"
+#define LANTERN_DEFAULT_VALIDATOR_REGISTRY "./genesis/validators.yaml"
+#define LANTERN_DEFAULT_NODES_FILE "./genesis/nodes.yaml"
+#define LANTERN_DEFAULT_GENESIS_STATE "./genesis/genesis.ssz"
+#define LANTERN_DEFAULT_VALIDATOR_CONFIG "./genesis/validator-config.yaml"
+#define LANTERN_DEFAULT_NODE_ID "lantern_0"
+#define LANTERN_DEFAULT_LISTEN_ADDR "/ip4/0.0.0.0/udp/9000/quic-v1"
+#define LANTERN_DEFAULT_HTTP_PORT 5052
+#define LANTERN_DEFAULT_METRICS_PORT 8080
+#define LANTERN_DEFAULT_DEVNET "devnet"
+
+struct lantern_client_options {
+    const char *data_dir;
+    const char *genesis_config_path;
+    const char *validator_registry_path;
+    const char *nodes_path;
+    const char *genesis_state_path;
+    const char *validator_config_path;
+    const char *node_id;
+    const char *node_key_hex;
+    const char *node_key_path;
+    const char *listen_address;
+    uint16_t http_port;
+    uint16_t metrics_port;
+    const char *devnet;
+    struct lantern_string_list bootnodes;
+    const char *hash_sig_key_dir;
+    const char *hash_sig_public_path;
+    const char *hash_sig_secret_path;
+    const char *hash_sig_public_template;
+    const char *hash_sig_secret_template;
+};
+
+void lantern_client_options_init(struct lantern_client_options *options);
+void lantern_client_options_free(struct lantern_client_options *options);
+int lantern_client_options_add_bootnode(struct lantern_client_options *options, const char *bootnode);
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif /* LANTERN_CORE_CLIENT_OPTIONS_H */
