+++ src/core/client/options.c
@@ -1,0 +1,43 @@
+#include "lantern/core/client/options.h"
+
+#include "lantern/support/string_list.h"
+
+void lantern_client_options_init(struct lantern_client_options *options) {
+    if (!options) {
+        return;
+    }
+
+    options->data_dir = LANTERN_DEFAULT_DATA_DIR;
+    options->genesis_config_path = LANTERN_DEFAULT_GENESIS_CONFIG;
+    options->validator_registry_path = LANTERN_DEFAULT_VALIDATOR_REGISTRY;
+    options->nodes_path = LANTERN_DEFAULT_NODES_FILE;
+    options->genesis_state_path = LANTERN_DEFAULT_GENESIS_STATE;
+    options->validator_config_path = LANTERN_DEFAULT_VALIDATOR_CONFIG;
+    options->node_id = LANTERN_DEFAULT_NODE_ID;
+    options->node_key_hex = NULL;
+    options->node_key_path = NULL;
+    options->listen_address = LANTERN_DEFAULT_LISTEN_ADDR;
+    options->http_port = LANTERN_DEFAULT_HTTP_PORT;
+    options->metrics_port = LANTERN_DEFAULT_METRICS_PORT;
+    options->devnet = LANTERN_DEFAULT_DEVNET;
+    lantern_string_list_init(&options->bootnodes);
+    options->hash_sig_key_dir = NULL;
+    options->hash_sig_public_path = NULL;
+    options->hash_sig_secret_path = NULL;
+    options->hash_sig_public_template = NULL;
+    options->hash_sig_secret_template = NULL;
+}
+
+void lantern_client_options_free(struct lantern_client_options *options) {
+    if (!options) {
+        return;
+    }
+    lantern_string_list_reset(&options->bootnodes);
+}
+
+int lantern_client_options_add_bootnode(struct lantern_client_options *options, const char *bootnode) {
+    if (!options || !bootnode) {
+        return -1;
+    }
+    return lantern_string_list_append(&options->bootnodes, bootnode);
+}
